update the command line help/description messages
