+++ client/attach.go
@@ -11,7 +11,7 @@ import (
 
 func (cli *HyperClient) HyperCmdAttach(args ...string) error {
 	var parser = gflag.NewParser(nil, gflag.Default)
-	parser.Usage = "attach CONTAINER\n\nAttach to the tty of a specified container in a pod"
+	parser.Usage = "attach CONTAINER\n\nAttach to the input/output of a specified container"
 	args, err := parser.ParseArgs(args)
 	if err != nil {
 		if !strings.Contains(err.Error(), "Usage") {
