+++ client/create.go
@@ -4,26 +4,48 @@ import (
 	"encoding/json"
 	"fmt"
 	"net/http"
+	"os"
+	"strings"
+
+	gflag "github.com/jessevdk/go-flags"
 
 	apitype "github.com/hyperhq/hyperd/types"
 )
 
 func (cli *HyperClient) HyperCmdCreate(args ...string) error {
-	copt, err := cli.ParseCreateOptions("create", args...)
-	if copt == nil {
+	var (
+		parser *gflag.Parser
+		opts   = &CreateFlags{}
+		err    error
+		podId  string
+	)
+	parser = gflag.NewParser(opts, gflag.Default|gflag.IgnoreUnknown|gflag.PassAfterNonOption)
+	parser.Usage = "create [OPTIONS] [POD_ID] IMAGE [COMMAND] [ARG...]\n\nCreate a pod, or create a container in the pod specified by the POD_ID"
+	args, err = parser.ParseArgs(args)
+	if err != nil {
+		if !strings.Contains(err.Error(), "Usage") {
+			return err
+		} else {
+			return nil
+		}
+	}
+
+	if opts.Container {
+		if len(args) == 0 {
+			return fmt.Errorf("%s: \"create\" requires the pod ID as first argument.", os.Args[0])
+		}
+		podId = args[0]
+		args = args[1:]
+	}
+
+	specjson, err := cli.ParseCommonOptions(&opts.CommonFlags, opts.Container, args...)
+	if err != nil {
 		return err
 	}
-	if copt.Remove {
-		return fmt.Errorf("\"create\" does not support rm parameter")
-	}
 
-	if copt.IsContainer && copt.PodId == "" {
-		return fmt.Errorf("did not provide the target pod")
-	}
-
-	if !copt.IsContainer {
+	if !opts.Container {
 		var tmpPod apitype.UserPod
-		err := json.Unmarshal(copt.JsonBytes, &tmpPod)
+		err := json.Unmarshal(specjson, &tmpPod)
 		if err != nil {
 			return fmt.Errorf("failed to read json: %v", err)
 		}
@@ -43,18 +65,18 @@ func (cli *HyperClient) HyperCmdCreate(args ...string) error {
 		fmt.Printf("Pod ID is %s\n", podId)
 	} else {
 		var tmpContainer apitype.UserContainer
-		err := json.Unmarshal(copt.JsonBytes, &tmpContainer)
+		err := json.Unmarshal(specjson, &tmpContainer)
 		if err != nil {
 			return fmt.Errorf("failed to read json: %v", err)
 		}
-		cid, statusCode, err := cli.client.CreateContainer(copt.PodId, &tmpContainer)
+		cid, statusCode, err := cli.client.CreateContainer(podId, &tmpContainer)
 		if err != nil {
 			if statusCode == http.StatusNotFound {
 				err = cli.PullImage(tmpContainer.Image)
 				if err != nil {
 					return err
 				}
-				cid, statusCode, err = cli.client.CreateContainer(copt.PodId, &tmpContainer)
+				cid, statusCode, err = cli.client.CreateContainer(podId, &tmpContainer)
 			}
 			if err != nil {
 				return err
