+++ client/exec.go
@@ -17,7 +17,7 @@ func (cli *HyperClient) HyperCmdExec(args ...string) error {
 		VM     bool `short:"m" long:"vm" description:"Execute outside of any containers"`
 	}
 	var parser = gflag.NewParser(&opts, gflag.Default|gflag.IgnoreUnknown)
-	parser.Usage = "exec [OPTIONS] POD|CONTAINER COMMAND [ARGS...]\n\nRun a command in a container of a running pod"
+	parser.Usage = "exec [OPTIONS] POD|CONTAINER COMMAND [ARGS...]\n\nRun a command in a container or a Pod"
 	args, err := parser.ParseArgs(args)
 	if err != nil {
 		if !strings.Contains(err.Error(), "Usage") {
